+++ CHANGELOG.md
@@ -29,6 +29,7 @@ ClimateDT workflow modifications:
 - New position for push_analysis `config.grouping.yaml` in `$AQUA/config/analysis`
 
 AQUA core complete list:
+- Centralise `save_figure` in OutputSaver class (#2425)
 - Allow longer time default for connection to dask cluster (#2420)
 - Safe parallel creation of area and weight files (#2412)
 - New grid lat-lon-r100 with explicit grid file, new default in DROP catgen (#2410, #2413)